Refactor index type from int32 to int16 for memory optimization
Replaced all occurrences of int32 indices with int16 to reduce memory usage and enhance performance where applicable. Removed unused wave displacement computation logic in vertex generation.

+++ MightyRenderingGuide/Source/MightyRenderingGuideGfx/Private/DynamicMesh/MrgSceneProxy.h
@@ -14,8 +14,7 @@ public:
 
 	virtual void InitRHI(FRHICommandListBase& RHICmdList) override;
 
-	// private:
-	TArray<int32> Indices;
+	TArray<int16> Indices;
 };
 
 class FMrgSceneProxy : public FPrimitiveSceneProxy
@@ -49,5 +48,5 @@ private:
 
 	FMrgVertexFactory VertexFactory;
 	TArray<FDynamicMeshVertex> Vertices;
-	TArray<int32> Indices;
+	TArray<int16> Indices;
 };
